refactor(realm): extract serial bridge from js-realm-shared
